requirements.txt
web3
py-geth
py-solc
coloredlogs
pycryptodome